requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==4.2.7
django-cors-headers==4.3.1
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
